// client/src/app/_models/member.ts
import {Photo} from "./photo";

export interface Member {
  id: number
  userName: string
  age: number
  photoUrl: string
  knownAs: string
  created:  Date
  lasActive: Date
  gender: string
  introduction: string
  interests: string
  lookingFor: string
  city: string
  country: string
  photos: Photo[]
}